refactoring: extract tag synchronization to extra class
