+++ src/test/java/de/unistuttgart/iste/gits/content_service/service/TagSynchronizerTest.java
@@ -1,0 +1,129 @@
+package de.unistuttgart.iste.gits.content_service.service;
+
+import de.unistuttgart.iste.gits.content_service.persistence.dao.ContentEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.dao.TagEntity;
+import de.unistuttgart.iste.gits.content_service.persistence.repository.TagRepository;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+@ExtendWith(MockitoExtension.class)
+public class TagSynchronizerTest {
+    @Mock
+    private TagRepository tagRepository;
+    @InjectMocks
+    private TagSynchronizer tagSynchronizer;
+
+    Set<String> getTagNames(Collection<TagEntity> tags) {
+        return tags.stream().filter(Objects::nonNull).map(TagEntity::getName).collect(Collectors.toSet());
+    }
+
+    @Test
+    void testTagSynchronisationNoExistingTagsSuccessful() {
+        List<String> tagNames = List.of("Tag1", "Tag2", "Tag3");
+        TagSynchronizer.TagSynchronizationResult result = tagSynchronizer.prepareSynchronization(tagNames,
+                Collections.emptyList(), Collections.emptyList());
+        assertThat(getTagNames(result.newTagsToAdd()), containsInAnyOrder(tagNames.toArray()));
+    }
+
+    @Test
+    void testTagSynchronisationWithExistingTagsSuccessful() {
+        List<String> tagNames = List.of("Tag1", "Tag2", "Tag3");
+        TagEntity existingTag = TagEntity.fromName(tagNames.get(1));
+        List<String> newTagNamesToAdd = List.of(tagNames.get(0), tagNames.get(2));
+        TagSynchronizer.TagSynchronizationResult result = tagSynchronizer.prepareSynchronization(tagNames,
+                Collections.emptyList(),
+                List.of(existingTag));
+        assertThat(getTagNames(result.newTagsToAdd()), containsInAnyOrder(newTagNamesToAdd.toArray()));
+        assertThat(getTagNames(result.existingTagsToAdd()), containsInAnyOrder(List.of(existingTag.getName()).toArray()));
+    }
+
+    @Test
+    void testTagSynchronisationWithAlreadyAssignedTagSuccessful() {
+        List<String> tagNames = List.of("Tag1", "Tag2", "Tag3");
+        TagEntity existingTag = TagEntity.fromName(tagNames.get(1));
+        List<String> newTagNamesToAdd = List.of(tagNames.get(0), tagNames.get(2));
+        TagSynchronizer.TagSynchronizationResult result = tagSynchronizer.prepareSynchronization(tagNames,
+                List.of(existingTag),
+                List.of(existingTag));
+        assertThat(getTagNames(result.newTagsToAdd()), containsInAnyOrder(newTagNamesToAdd.toArray()));
+        assertThat(getTagNames(result.existingTagsToAdd()), containsInAnyOrder(Collections.emptyList().toArray()));
+        assertThat(getTagNames(result.existingTagsToRemove()), containsInAnyOrder(Collections.emptyList().toArray()));
+    }
+
+    @Test
+    void testTagSynchronisationWithTagsToRemoveSuccessful() {
+        List<String> tagNames = List.of("Tag1");
+        TagEntity existingTag = TagEntity.fromName("Tag2");
+        List<String> newTagNamesToAdd = List.of(tagNames.get(0));
+        List<String> existingTagNamesToRemove = List.of("Tag2");
+        TagSynchronizer.TagSynchronizationResult result = tagSynchronizer.prepareSynchronization(tagNames,
+                List.of(existingTag),
+                List.of(existingTag));
+        assertThat(getTagNames(result.newTagsToAdd()), containsInAnyOrder(newTagNamesToAdd.toArray()));
+        assertThat(getTagNames(result.existingTagsToAdd()), containsInAnyOrder(Collections.emptyList().toArray()));
+        assertThat(getTagNames(result.existingTagsToRemove()), containsInAnyOrder(existingTagNamesToRemove.toArray()));
+    }
+
+    @Test
+    void testSynchronizeWithDbAddNewTagsSuccessful() {
+        List<String> tagNames = List.of("Tag1");
+        TagEntity newTag = TagEntity.fromName(tagNames.get(0));
+        TagSynchronizer.TagSynchronizationResult preparation = new TagSynchronizer.TagSynchronizationResult(
+                Collections.emptyList(),
+                List.of(newTag),
+                Collections.emptyList()
+        );
+        ContentEntity content = new ContentEntity();
+        content.setId(UUID.randomUUID());
+        when(tagRepository.save(Mockito.any(TagEntity.class))).thenAnswer(i -> i.getArguments()[0]);
+        tagSynchronizer.synchronizeWithDatabase(content, preparation);
+        verify(tagRepository).save(newTag);
+        assertThat(content.getTags(), is(equalTo(Set.of(newTag))));
+        assertThat(newTag.getContents(), is(equalTo(Set.of(content))));
+    }
+
+    @Test
+    void testSynchronizeWithDbAddNewContentToExistingTagSuccessful() {
+        List<String> tagNames = List.of("Tag1");
+        TagEntity existingTag = TagEntity.fromName(tagNames.get(0));
+        TagSynchronizer.TagSynchronizationResult preparation = new TagSynchronizer.TagSynchronizationResult(
+                List.of(existingTag),
+                Collections.emptyList(),
+                Collections.emptyList()
+        );
+        ContentEntity content = new ContentEntity();
+        content.setId(UUID.randomUUID());
+        tagSynchronizer.synchronizeWithDatabase(content, preparation);
+        assertThat(content.getTags(), is(equalTo(Set.of(existingTag))));
+        assertThat(existingTag.getContents(), is(equalTo(Set.of(content))));
+    }
+
+    @Test
+    void testSynchronizeWithDbRemoveContentFromExistingTagSuccessful() {
+        ContentEntity content = new ContentEntity();
+        content.setId(UUID.randomUUID());
+        TagEntity existingTag = TagEntity.fromName("Tag1");
+        content.addToTags(existingTag);
+        existingTag.addToContents(content);
+        TagSynchronizer.TagSynchronizationResult preparation = new TagSynchronizer.TagSynchronizationResult(
+                Collections.emptyList(),
+                Collections.emptyList(),
+                List.of(existingTag)
+        );
+        tagSynchronizer.synchronizeWithDatabase(content, preparation);
+        assertThat(content.getTags().size(), is(equalTo(0)));
+        assertThat(existingTag.getContents().size(), is(equalTo(0)));
+    }
+}
